racing to the end of the universe
Starting to add the graphics to the game!

added a moving background within the playscreen (moving road)
added a background image for the main page
added a background image for the highscore page
rather then checking if the car is moved i need to check whether the car has been moved left and right, as different images will be displayed when it turns. This was done and checked with print statements, but have since been taken out as it is working (as shown by drawing the player with different colours )

loading the images has error correcting, however this is only at the start and the game will later fail as it wont be able to use the variables.
Highscore screen will be done last as i will use a screen shot from the finished play screen.